\name{IM-internal}
\alias{Numbers-class}
\alias{initialize,MultiImages-method}
\alias{checkSize<-}
\alias{Invariant<-}
\alias{setPoly<-}
\alias{Moments<-}
\alias{geoMoment}
\alias{getPAxis}
\alias{vlm}
\alias{polarize}
\alias{revPolarize}
\alias{binarize}
\alias{getTheta}
\alias{momentType<-}
\alias{setCentroid<-}
\alias{Reconstruct<-}
\alias{setParams<-}
\alias{setOrder<-}
\alias{setPolar<-}
\alias{transform<-}
\alias{plotPoly}
\alias{setImage<-}
\alias{scaleTC}
\alias{OrthReconstruct}
\alias{OrthPolynomials}
\alias{OrthMoments}
\alias{mapCenter}
\alias{Legendre}
\alias{KrawWeights}
\alias{Krawtchouk}
\alias{GPZMreconMulti}
\alias{GPZMrecon}
\alias{GPZMmultiple}
\alias{GPZMmoments}
\alias{Gegenbauer}
\alias{FRrecon}
\alias{FRmoments}
\alias{FMrecon}
\alias{FMmoments}
\alias{FCrecon}
\alias{FCmoments}
\alias{DualHahn}
\alias{CmplxreconMulti}
\alias{Cmplxmultiple}
\alias{ChebyshevCont}
\alias{Chebyshev}
\alias{flip}
\alias{zeroMoment}
\title{Internal Image Moment Functions}
\description{
          Internal IM functions
}
\details{
          These are not to be called by the user.
}
\keyword{internal}